+++ ProdManager/models/Maintenance.py
@@ -1,5 +1,3 @@
-import enum
-
 from sqlalchemy import String, Integer, Column, ForeignKey, DateTime, Enum
 from sqlalchemy.orm import relationship
 
@@ -10,11 +8,9 @@ from .Service import ServiceStatus
 from .MaintenanceComment import MaintenanceComment
 
 class MaintenanceStatus(ModelEnum):
-  CREATED = 'created'
-  VALIDATED = 'validated'
-  SHEDULED = 'sheduled'
+  SCHEDULED = 'scheduled'
   IN_PROGRESS = 'in-progress'
-  SUCCESS = 'success'
+  SUCCEED = 'succeed'
   FAILED = 'failed'
 
 class Maintenance(db.Model):
@@ -25,23 +21,29 @@ class Maintenance(db.Model):
   status = Column(Enum(MaintenanceStatus), nullable=False)
   scope_id = Column(Integer, ForeignKey('scope.id'), nullable=False)
   service_id = Column(Integer, ForeignKey('service.id'), nullable=False)
-  service_planned_status = Column(Enum(ServiceStatus), nullable=False)
+  service_status = Column(Enum(ServiceStatus), nullable=False)
   creation_date = Column(DateTime(), nullable=False)
-  scheduled_start_date = Column(DateTime(), nullable=True)
-  scheduled_end_date = Column(DateTime(), nullable=True)
+  scheduled_start_date = Column(DateTime(), nullable=False)
+  scheduled_end_date = Column(DateTime(), nullable=False)
   start_date = Column(DateTime(), nullable=True)
   end_date = Column(DateTime(), nullable=True)
-  comments = relationship('MaintenanceComment', backref='maintenance', lazy='dynamic', order_by='desc(MaintenanceComment.creation_date)', cascade="all, delete")
+  comments = relationship(
+    'MaintenanceComment',
+    backref='maintenance',
+    lazy='dynamic',
+    order_by='desc(MaintenanceComment.creation_date)',
+    cascade="all, delete"
+)
 
   def __repr__(self):
     return f"<Maintenance '{self.name}'>"
 
 def filter_ongoing_maintenance(query, limit=10):
   return query.filter(
-    Maintenance.status.in_([MaintenanceStatus.SHEDULED, MaintenanceStatus.IN_PROGRESS])
+    Maintenance.status.in_([MaintenanceStatus.SCHEDULED, MaintenanceStatus.IN_PROGRESS])
   ).limit(limit)
 
 def filter_past_maintenance(query, limit=10):
   return query.filter(
-    Maintenance.status.in_([MaintenanceStatus.SUCCESS, MaintenanceStatus.FAILED])
+    Maintenance.status.in_([MaintenanceStatus.SUCCEED, MaintenanceStatus.FAILED])
   ).limit(limit)