+++ ProdManager/routes/incident/forms.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from markupsafe import escape
 
 from flask_wtf import FlaskForm
 from wtforms import StringField, TextAreaField, SelectField, DateTimeLocalField
@@ -8,39 +7,48 @@ from wtforms.validators import DataRequired, Length, Optional
 from ProdManager.models.Incident import IncidentSeverity, IncidentStatus
 
 class IncidentCreateForm(FlaskForm):
+  scope_id = SelectField(name='scope', validators=[DataRequired()])
+  service_id = SelectField(name='service', validators=[DataRequired()])
   name = StringField(name='name', validators=[DataRequired(), Length(min=4)])
+  description = TextAreaField(name='description', validators=[Optional()])
   external_reference= StringField(name='external_reference', validators=[Optional()])
-  severity = SelectField(name='severity', validators=[DataRequired()], choices=IncidentSeverity.choices(), coerce=IncidentSeverity.coerce)
+  severity = SelectField(
+    name='severity',
+    validators=[DataRequired()],
+    choices=IncidentSeverity.choices(),
+    coerce=IncidentSeverity.coerce,
+  )
   start_impact_date = DateTimeLocalField(
     name='start impact date',
     default=datetime.now(),
     validators=[Optional()],
-    format='%Y-%m-%dT%H:%M'
+    format='%Y-%m-%dT%H:%M',
   )
-  description = TextAreaField(name='description', validators=[Optional()])
-  scope_id = SelectField(name='scope', validators=[DataRequired()])
-  service_id = SelectField(name='service', validators=[DataRequired()])
 
 class IncidentUpdateForm(FlaskForm):
+  scope_id = SelectField(name='scope', validators=[DataRequired()])
+  service_id = SelectField(name='service', validators=[DataRequired()])
   name = StringField(name='name', validators=[DataRequired(), Length(min=4)])
+  description = TextAreaField(name='description', validators=[Optional()])
   external_reference= StringField(name='external_reference', validators=[Optional()])
-  severity = SelectField(name='severity', validators=[DataRequired()], choices=IncidentSeverity.choices(), coerce=IncidentSeverity.coerce)
-  status = SelectField(name='status', validators=[DataRequired()], choices=IncidentStatus.choices(), coerce=IncidentStatus.coerce)
+  severity = SelectField(
+    name='severity',
+    validators=[DataRequired()],
+    choices=IncidentSeverity.choices(),
+    coerce=IncidentSeverity.coerce,
+    )
+  status = SelectField(
+    name='status',
+    validators=[DataRequired()],
+    choices=IncidentStatus.choices(),
+    coerce=IncidentStatus.coerce,
+    )
   start_impact_date = DateTimeLocalField(
     name='start impact date',
     default=datetime.now(),
     validators=[Optional()],
-    format='%Y-%m-%dT%H:%M'
+    format='%Y-%m-%dT%H:%M',
   )
-  resolve_date = DateTimeLocalField(
-    name='resolve date',
-    default=datetime.now(),
-    validators=[Optional()],
-    format='%Y-%m-%dT%H:%M'
-  )
-  description = TextAreaField(name='description', validators=[Optional()])
-  scope_id = SelectField(name='scope', validators=[DataRequired()])
-  service_id = SelectField(name='service', validators=[DataRequired()])
 
 class IncidentDeleteForm(FlaskForm):
   pass