+++ ProdManager/routes/maintenance/forms.py
@@ -9,55 +9,60 @@ from ProdManager.models.Maintenance import MaintenanceStatus
 from ProdManager.models.Service import ServiceStatus
 
 class MaintenanceCreateForm(FlaskForm):
+  scope_id = SelectField(name='scope', validators=[DataRequired()])
+  service_id = SelectField(name='service', validators=[DataRequired()])
   name = StringField(name='name', validators=[DataRequired(), Length(min=4)])
+  description = TextAreaField(name='description', validators=[Optional()])
   external_reference= StringField(name='external_reference', validators=[Optional()])
+  service_status = SelectField(
+    name='service status',
+    validators=[DataRequired()],
+    choices=ServiceStatus.choices(),
+    coerce=ServiceStatus.coerce,
+  )
   scheduled_start_date = DateTimeLocalField(
     name='scheduled start date',
     default=datetime.now(),
-    validators=[Optional()],
-    format='%Y-%m-%dT%H:%M'
+    validators=[DataRequired()],
+    format='%Y-%m-%dT%H:%M',
   )
   scheduled_end_date = DateTimeLocalField(
     name='scheduled end date',
     default=datetime.now() + timedelta(hours=1),
-    validators=[Optional()],
-    format='%Y-%m-%dT%H:%M'
+    validators=[DataRequired()],
+    format='%Y-%m-%dT%H:%M',
   )
-  service_planned_status = SelectField(name='service planed status', validators=[DataRequired()], choices=ServiceStatus.choices(), coerce=ServiceStatus.coerce)
-  description = TextAreaField(name='description', validators=[Optional()])
-  scope_id = SelectField(name='scope', validators=[DataRequired()])
-  service_id = SelectField(name='service', validators=[DataRequired()])
 
 class MaintenanceUpdateForm(FlaskForm):
+  scope_id = SelectField(name='scope', validators=[DataRequired()])
+  service_id = SelectField(name='service', validators=[DataRequired()])
   name = StringField(name='name', validators=[DataRequired(), Length(min=4)])
+  description = TextAreaField(name='description', validators=[Optional()])
   external_reference= StringField(name='external_reference', validators=[Optional()])
-  status = SelectField(name='status', validators=[DataRequired()], choices=MaintenanceStatus.choices(), coerce=MaintenanceStatus.coerce)
+  status = SelectField(
+    name='status',
+    validators=[DataRequired()],
+    choices=MaintenanceStatus.choices(),
+    coerce=MaintenanceStatus.coerce,
+  )
+  service_status = SelectField(
+    name='service status',
+    validators=[DataRequired()],
+    choices=ServiceStatus.choices(),
+    coerce=ServiceStatus.coerce,
+  )
   scheduled_start_date = DateTimeLocalField(
     name='scheduled start date',
     default=datetime.now(),
-    validators=[Optional()],
-    format='%Y-%m-%dT%H:%M'
+    validators=[DataRequired()],
+    format='%Y-%m-%dT%H:%M',
   )
   scheduled_end_date = DateTimeLocalField(
     name='scheduled end date',
     default=datetime.now() + timedelta(hours=1),
-    validators=[Optional()],
-    format='%Y-%m-%dT%H:%M'
+    validators=[DataRequired()],
+    format='%Y-%m-%dT%H:%M',
   )
-  start_date = DateTimeLocalField(
-    name='start date',
-    validators=[Optional()],
-    format='%Y-%m-%dT%H:%M'
-  )
-  end_date = DateTimeLocalField(
-    name='end date',
-    validators=[Optional()],
-    format='%Y-%m-%dT%H:%M'
-  )
-  service_planned_status = SelectField(name='service planed status', validators=[DataRequired()], choices=ServiceStatus.choices(), coerce=ServiceStatus.coerce)
-  description = TextAreaField(name='description', validators=[Optional()])
-  scope_id = SelectField(name='scope', validators=[DataRequired()])
-  service_id = SelectField(name='service', validators=[DataRequired()])
 
 class MaintenanceDeleteForm(FlaskForm):
   pass