+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 
 ## [Unreleased]
 
+### Added
+
+- feat(template): Add new filter 'format_column_name' to beautify the column name for tables
+- feat(template.Service): Display service status
+- feat(view.Service): Allow to update service status
+- feat(Template.Timeline): Introduce timeline display
+- feat(Model.Incident): Add new columns 'investigation_date', 'stable_date'
+- feat(Model.Incident): Add 'INVESTIGATING' status
+- feat(Style.Incident): Improve incident display
+- feat(Template.Incident): Improve incident display
+- feat(Style.Maintenance): Improve maintenance display
+- feat(Template.Maintenance): Improve maintenance display
+
+### Changed
+
+- refactor(Model.Maintenance): Change column name from 'service_planned_status' into 'service_status'
+- change(template.\*): Remove description title
+- change(Model.Maintenance): Rename 'SHEDULED' status in 'SCHEDULED'
+
+### Fixed
+
+- fix(\*.Maintenance): Don't use deprecated 'service_planned_status' attribute
+
+### Breaking Change
+
+- change(Model.Incident): Remove 'COMPLETED' status
+- change(Model.Maintenance): Remove 'CREATED', 'VALIDATED' status
 
 ## 0.2.0